% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_question.R
\name{see_question}
\alias{see_question}
\alias{see_question.data.frame}
\alias{see_question.numeric}
\title{View census form for particular question}
\usage{
see_question(...)

\method{see_question}{data.frame}(.data, ...)

\method{see_question}{numeric}(qn, ...)
}
\arguments{
\item{...}{Passed to method: either a question number or a data frame.}

\item{.data}{A \code{data.frame} from this package.}

\item{qn}{Question number.}
}
\value{
Primarily called for its side effect: plots an image of the question (and available answers, if appropriate) in the plot window.
If a \code{data.frame} from this package is called, it is returned (invisibly if \code{knitr.in.progress} is \code{TRUE}).
}
\description{
View census form for particular question
}
\examples{
\dontrun{
see_question(5)
see_question(Census2016_languages)
}

}
