% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Census2016_2D_tables.R
\docType{data}
\name{TwoDimTbl}
\alias{TwoDimTbl}
\alias{Census2016_ancestories}
\alias{Census2016_countries_of_birth}
\alias{Census2016_languages}
\alias{Census2016_religions}
\title{Two dimensional tables}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 201600 rows and 6 columns.}
\usage{
Census2016_ancestories

Census2016_countries_of_birth

Census2016_languages

Census2016_religions
}
\description{
Tables with two keys (the SA2 and a label for which the count applies)
}
\details{
From the Community Series data packs. 
N.B. \code{countries_of_birth} only includes the top 35 responses countries of birth across Australia.
}
\keyword{datasets}
